Added failing test for 'after all' not skipped by tags
