+++ NSpecSpecs/NSpecSpecs.csproj
@@ -89,7 +89,7 @@
     <Compile Include="describe_RunningSpecs\describe_before_and_after\async_class_levels.cs" />
     <Compile Include="describe_RunningSpecs\describe_before_and_after\async_when_there_are_no_specs.cs" />
     <Compile Include="describe_RunningSpecs\describe_method_level_afters.cs" />
-    <Compile Include="describe_RunningSpecs\describe_skipped_before_alls_when_excluded_by_tag.cs" />
+    <Compile Include="describe_RunningSpecs\describe_skipped_before_after_alls_when_excluded_by_tag.cs" />
     <Compile Include="describe_RunningSpecs\describe_before_and_after\abstract_class.cs" />
     <Compile Include="describe_RunningSpecs\describe_before_and_after\nested_contexts.cs" />
     <Compile Include="describe_RunningSpecs\describe_before_and_after\when_there_are_no_specs.cs" />
